# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/discovery/preferences": {
      "get": {
        "operationId": "Discovery_getPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get the current discovery preference filters",
        "tags": [
          "discovery"
        ]
      },
      "patch": {
        "operationId": "Discovery_updatePreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update discovery preference filters and reset queue lineage",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/feed": {
      "get": {
        "operationId": "Discovery_getFeed",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Fetch a discovery feed window using the current queue cursor",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/window": {
      "get": {
        "operationId": "Discovery_getWindow",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Fetch the authoritative discovery deck window",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/reset": {
      "post": {
        "operationId": "Discovery_resetDecisions",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Clear discovery decisions for the current actor",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/like": {
      "post": {
        "operationId": "Discovery_likeProfile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Like a discovery profile and request the next queue state",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/decision": {
      "post": {
        "operationId": "Discovery_decideProfile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Apply one queued like or pass decision",
        "tags": [
          "discovery"
        ]
      }
    },
    "/api/discovery/pass": {
      "post": {
        "operationId": "Discovery_passProfile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Pass on a discovery profile and request the next queue state",
        "tags": [
          "discovery"
        ]
      }
    },
...
    "/api/me/discovery/preferences": {
      "get": {
        "operationId": "Viewer_getDiscoveryPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Get discovery preferences from the viewer surface",
        "tags": [
          "viewer"
        ]
      },
      "patch": {
        "operationId": "Viewer_updateDiscoveryPreferences",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "matcha-bearer": []
          }
        ],
        "summary": "Update discovery preferences from the viewer surface",
        "tags": [
          "viewer"
        ]
      }
    }
  },
  "info": {
    "title": "Matcha API",
    "description": "Canonical API contract for the Matcha backend. Target OpenAPI 3.1 where the active toolchain remains smooth.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "matcha-bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Bearer access token returned by Matcha authentication endpoints."
      }
    },
...
  }
}